hhparser/telegram_bot/requirements.txt
python-telegram-bot==20.0
requests==2.28.1
psycopg2-binary==2.9.5
pandas==1.5.3
numpy==1.23.5
python-dotenv==0.19.1